+++ packages/react-ui/components/RadioGroup/RadioGroup.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { AriaAttributes } from 'react';
 import invariant from 'invariant';
 
 import { getRandomID, isNonNullable } from '../../lib/utils';
@@ -74,6 +74,10 @@ export interface RadioGroupProps<T = string | number> extends CommonProps {
    * Принимает два аргумента: `(value: Value, data: Data) => React.Node`
    */
   renderItem?: (itemValue: T, data: React.ReactNode) => React.ReactNode;
+  /**
+   * Атрибут для указания id элемента(-ов), описывающих его
+   */
+  'aria-describedby'?: AriaAttributes['aria-describedby'];
   /** Вызывается при изменении `value` */
   onValueChange?: (value: T) => void;
   onBlur?: (event: FocusEvent) => void;
@@ -149,7 +153,7 @@ export class RadioGroup<T> extends React.Component<RadioGroupProps<T>, RadioGrou
   }
 
   public renderMain() {
-    const { width, onMouseLeave, onMouseOver, onMouseEnter, onBlur } = this.props;
+    const { width, onMouseLeave, onMouseOver, onMouseEnter, onBlur, 'aria-describedby': ariaDescribedby } = this.props;
     const style = {
       width: width ?? 'auto',
     };
@@ -162,7 +166,15 @@ export class RadioGroup<T> extends React.Component<RadioGroupProps<T>, RadioGrou
     return (
       <CommonWrapper rootNodeRef={this.setRootNode} {...this.props}>
         <FocusTrap onBlur={onBlur}>
-          <span data-tid={RadioGroupDataTids.root} ref={this.ref} style={style} className={styles.root()} {...handlers}>
+          <span
+            data-tid={RadioGroupDataTids.root}
+            ref={this.ref}
+            style={style}
+            className={styles.root()}
+            role="radiogroup"
+            {...handlers}
+            aria-describedby={ariaDescribedby}
+          >
             <RadioGroupContext.Provider value={this.getRadioGroupContextValue()}>
               {this.renderChildren()}
             </RadioGroupContext.Provider>
@@ -225,7 +237,7 @@ export class RadioGroup<T> extends React.Component<RadioGroupProps<T>, RadioGrou
     };
 
     return (
-      <span {...itemProps}>
+      <span {...itemProps} role="presentation">
         <Radio value={itemValue}>{this.getProps().renderItem(itemValue, data)}</Radio>
       </span>
     );